Fix `test/jest`, `shared/modules/selectors`
