+++ shared/modules/selectors/util.ts
@@ -1,4 +1,7 @@
-import { TransactionStatus } from '@metamask/transaction-controller';
+import {
+  TransactionMeta,
+  TransactionStatus,
+} from '@metamask/transaction-controller';
 import { isEqual } from 'lodash';
 import { createSelectorCreator, defaultMemoize } from 'reselect';
 
@@ -7,11 +10,13 @@ export const createDeepEqualSelector = createSelectorCreator(
   isEqual,
 );
 
-export const filterAndShapeUnapprovedTransactions = (transactions) => {
+export const filterAndShapeUnapprovedTransactions = (
+  transactions: TransactionMeta[],
+) => {
   return transactions
     .filter(({ status }) => status === TransactionStatus.unapproved)
-    .reduce((result, transaction) => {
+    .reduce<Record<string, TransactionMeta>>((result, transaction) => {
       result[transaction.id] = transaction;
       return result;
-    }, {});
+    }, {} as never);
 };
